Use original filename from enclosure when given
Previously, all files came in with no extensions, practically requiring
you to disable extension validation to import files, and resulting in
files with no extension being saved.

This change uses the title attribute of the enclosure link to set the
original filename, if Zotero supplies it. Unfortunately, some
attachments (like snapshots) don't include the filename: for these, the
old system that simply uses the basename of the S3 URI is maintained.